+++ tachyon/math/finite_fields/fp3.h
@@ -41,7 +41,7 @@ class Fp3 final : public CubicExtensionField<Fp3<Config>> {
     //    = α₀ + α₁(x³)^((P - 1) / 3) * x + α₂(x³)^(2 * (P - 1) / 3) * x²
     //    = α₀ + α₁ωx + α₂ω²x², where ω is a cubic root of unity.
 
-    constexpr uint64_t N = BasePrimeField::kLimbNums;
+    constexpr size_t N = BasePrimeField::kLimbNums;
     // m₁ = P
     mpz_class m1;
     if constexpr (BasePrimeField::Config::kModulusBits <= 32) {
